Add files via upload
We, a team of four, collaborated effectively to design 3D models of oil refinery parts using AutoCAD. Our teamwork was evident in the seamless integration of our individual skills and knowledge. We successfully created a comprehensive presentation, highlighting the functionality and purpose of each part. The final presentation, complete with a poster and PPT, showcased our collective ability to visualize, design, and communicate complex engineering concepts.